backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
qdrant-client==1.11.0
google-generativeai==0.8.3
python-dotenv==1.0.1

# Authentication & Database (optional - only needed if using auth)
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9; python_version < "3.13"
psycopg2-binary>=2.9.10; python_version >= "3.13"
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
authlib==1.3.0

